FEATURE: luaossl: Add Lua bindings for OpenSSL
https://github.com/wahern/luaossl
http://25thandclement.com/~william/projects/luaossl.html

For now the module is linked with a static build of openssl to
prevent programs embedding the Lua interpreter from being tainted
by shared libraries (openssl) loaded by the module, which could
define alternate symbols than the ones originally linked with the
program.

The exact use case is related to problems we had in a build of
haproxy with support for the quic protocol. This requires a custom
openssl and we choose to link it statically when builting haproxy.
If the haproxy configuration makes reference to lua scripts that
load luaossl, when the later is link dynamically with the standard
openssl libs from the system, this creates symbol conflicts in
haproxy, leading to crashes.

The lua53z packaging may be subject to changes in future because
there is no reason to have only this module statically linked with
dependant libraires. For some modules like lua-curl, not depending
on the system standard shared libs would be too complex and too
hard to maintain.