BUILD: check-updates: Handle Github API pagination and ratelimit
When querying the Github API to list tags, the results are paginated.
This patch ensure we walk on all the pages so we do not miss a release
tag.

Requests made to the API are rate-limited to something like 60 per
hour. In order to push the limit to 5000 per hour, users can create an
access token and use it for Basic authentication. This is compatible
with curl's .netrc file, hence the use of the --netrc-optional curl
option.